Add files via upload
This is an updated correct version of the schema installation for orthoMCL file. For the reason that I didn't try running the file on other system, this is only safe to use under a Ubuntu system. The editor and IDE I used to modify the file is DataGrip from jetBrains. The sql codes uploaded are not original codes, but modified from the code downloaded from orthoMCL website.